refactoring: replaced layout segment into widgets.
